refactor: OAuth Response DTO 통일
