+++ src/main/java/mocacong/server/security/auth/OAuthPlatformMemberResponse.java
@@ -1,4 +1,4 @@
-package mocacong.server.security.auth.apple;
+package mocacong.server.security.auth;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class ApplePlatformMemberResponse {
+public class OAuthPlatformMemberResponse {
 
     private String platformId;
     private String email;
