+++ src/main/java/mocacong/server/security/auth/apple/AppleOAuthUserProvider.java
@@ -5,6 +5,7 @@ import java.security.PublicKey;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import mocacong.server.exception.unauthorized.InvalidTokenException;
+import mocacong.server.security.auth.OAuthPlatformMemberResponse;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -16,7 +17,7 @@ public class AppleOAuthUserProvider {
     private final PublicKeyGenerator publicKeyGenerator;
     private final AppleClaimsValidator appleClaimsValidator;
 
-    public ApplePlatformMemberResponse getApplePlatformMember(String identityToken) {
+    public OAuthPlatformMemberResponse getApplePlatformMember(String identityToken) {
         Map<String, String> headers = appleJwtParser.parseHeaders(identityToken);
         ApplePublicKeys applePublicKeys = appleClient.getApplePublicKeys();
 
@@ -24,7 +25,7 @@ public class AppleOAuthUserProvider {
 
         Claims claims = appleJwtParser.parsePublicKeyAndGetClaims(identityToken, publicKey);
         validateClaims(claims);
-        return new ApplePlatformMemberResponse(claims.getSubject(), claims.get("email", String.class));
+        return new OAuthPlatformMemberResponse(claims.getSubject(), claims.get("email", String.class));
     }
 
     private void validateClaims(Claims claims) {
